+++ src/main/java/blockswarm/network/cluster/Node.java
@@ -49,7 +49,7 @@ public class Node
     TrafficLimiter trafficLimiter = new TrafficLimiter();
     
     int TIMEOUT = 60 * 1000;
-    int PORT = 44446;
+    int PORT = 44448;
     
     public Node()
     {